Reviewing if User created post. Setting/Getting User Status.

+++ 25-REST-API/backend/controllers/feed.js
@@ -126,6 +126,13 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -157,11 +164,24 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      if (post.creator.toString() !== req.userId) {
+        console.log(post.creator.toString());
+        console.log(req.userId);
+        
+        const error = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
       // Check logged in user.
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     }).then(result => {
-      console.log(result);
+      return User.findById(req.userId)
+    }).then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    }).then(user => {
       res.status(200).json({ message: 'Deleted Post.'});
     }).catch(err => {
       if (!err.statusCode) {
